extract Mission components to missions directory
